// src/pages/Companies/schema.ts
import { DocumentSnapshot } from "@firebase/firestore";
import { CompanyCategory } from "../../types";

export interface PaginationState {
    pageIndex: number;
    pageSize: number;
    pageAction: "NEXT" | "PREV" | null;
    lastIndex: string | null;
    firstIndex: string | null;
    categoryFilterValue?: CompanyCategory;
    startAfterDoc: DocumentSnapshot | null;
}

export type ColumnDef =
    | {
        id: string;
        header: string | JSX.Element;
        canHide: true;
        isVisible: boolean;
    }
    | {
        id: string;
        header: string | JSX.Element;
        canHide: false;
        isVisible?: true;
    };